refactor: reduce cost of quantization virtual function call
