+++ crates/quantization/src/reranker/mod.rs
@@ -1,3 +1,2 @@
-pub mod disabled;
-pub mod window;
-pub mod window_0;
+pub mod flat;
+pub mod graph;
